+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/MatrixClient.kt
@@ -114,6 +114,8 @@ interface MatrixClient : AutoCloseable {
     suspend fun setDisplayName(displayName: String?): Result<Unit>
 
     suspend fun setAvatarUrl(avatarUrl: String?): Result<Unit>
+
+    suspend fun closeSuspending()
 }
 
 private val coroutineExceptionHandler = CoroutineExceptionHandler { _, exception ->
@@ -548,6 +550,7 @@ suspend fun MatrixClient.Companion.loginWith(
         eventContentSerializerMappings = di.get(),
         accountStore = accountStore,
         olmCryptoStore = di.get(),
+        coroutineContext = coroutineContext,
         config = config
     ) { matrixClient ->
         val keyStore = di.get<KeyStore>()
@@ -634,7 +637,8 @@ suspend fun MatrixClient.Companion.fromStore(
             eventContentSerializerMappings = di.get(),
             accountStore = accountStore,
             olmCryptoStore = olmCryptoStore,
-            config = config
+            config = config,
+            coroutineContext = coroutineContext,
         ) { matrixClient ->
             val accessToken = account.accessToken ?: onSoftLogin?.let {
                 val (identifier, password, token, loginType) = onSoftLogin()
@@ -661,10 +665,10 @@ private suspend fun initMatrixClientKoinApplication(
     config: MatrixClientConfiguration
 ): KoinApplication {
     val coroutineName = config.name?.let { name -> CoroutineName(name) }
-    val job = SupervisorJob(coroutineContext[Job])
-    val coroutineScope = CoroutineScope(coroutineContext + job + coroutineExceptionHandler).apply {
-        if (coroutineName != null) this + coroutineName
-    }
+    val coroutineScope =
+        CoroutineScope(coroutineContext + SupervisorJob(coroutineContext[Job]) + coroutineExceptionHandler).apply {
+            if (coroutineName != null) this + coroutineName
+        }
 
     val koinApplication = koinApplication {
         modules(module {
@@ -692,6 +696,7 @@ private suspend fun <T : MatrixClient?> KoinApplication.createMatrixClient(
     accountStore: AccountStore,
     olmCryptoStore: OlmCryptoStore,
     config: MatrixClientConfiguration,
+    coroutineContext: CoroutineContext,
     doFinally: suspend (MatrixClient) -> T,
 ): T {
     val api = config.matrixClientServerApiClientFactory.create(
@@ -712,7 +717,8 @@ private suspend fun <T : MatrixClient?> KoinApplication.createMatrixClient(
             }
         },
         syncLoopDelay = config.syncLoopDelays.syncLoopDelay,
-        syncLoopErrorDelay = config.syncLoopDelays.syncLoopErrorDelay
+        syncLoopErrorDelay = config.syncLoopDelays.syncLoopErrorDelay,
+        coroutineContext = coroutineContext,
     )
     val (signingKey, identityKey) = freeAfter(
         olmCryptoStore.getOlmAccount()
@@ -983,6 +989,12 @@ class MatrixClientImpl internal constructor(
         api.close()
         di.close()
         coroutineScope.cancel("stopped MatrixClient")
+    }
+
+    override suspend fun closeSuspending() {
+        val job = coroutineScope.coroutineContext.job
+        close()
+        job.join()
     }
 
     override suspend fun setDisplayName(displayName: String?): Result<Unit> {