+++ trixnity-clientserverapi/trixnity-clientserverapi-client/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/client/MatrixClientServerApiClient.kt
@@ -4,13 +4,12 @@ import io.github.oshai.kotlinlogging.KotlinLogging
 import io.ktor.client.*
 import io.ktor.client.engine.*
 import io.ktor.http.*
-import kotlinx.coroutines.CoroutineExceptionHandler
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.cancel
+import kotlinx.coroutines.*
 import kotlinx.serialization.json.Json
 import net.folivo.trixnity.core.serialization.createMatrixEventJson
 import net.folivo.trixnity.core.serialization.events.DefaultEventContentSerializerMappings
 import net.folivo.trixnity.core.serialization.events.EventContentSerializerMappings
+import kotlin.coroutines.CoroutineContext
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
 
@@ -48,6 +47,8 @@ interface MatrixClientServerApiClient : AutoCloseable {
 
     val eventContentSerializerMappings: EventContentSerializerMappings
     val json: Json
+
+    suspend fun closeSuspending()
 }
 
 interface MatrixClientServerApiClientFactory {
@@ -59,7 +60,7 @@ interface MatrixClientServerApiClientFactory {
         syncBatchTokenStore: SyncBatchTokenStore = SyncBatchTokenStore.inMemory(),
         syncLoopDelay: Duration = 2.seconds,
         syncLoopErrorDelay: Duration = 5.seconds,
-        syncCoroutineScope: CoroutineScope? = null,
+        coroutineContext: CoroutineContext = Dispatchers.Default,
         httpClientEngine: HttpClientEngine? = null,
         httpClientConfig: (HttpClientConfig<*>.() -> Unit)? = null,
     ): MatrixClientServerApiClient =
@@ -71,7 +72,7 @@ interface MatrixClientServerApiClientFactory {
             syncBatchTokenStore = syncBatchTokenStore,
             syncLoopDelay = syncLoopDelay,
             syncLoopErrorDelay = syncLoopErrorDelay,
-            syncCoroutineScope = syncCoroutineScope,
+            coroutineContext = coroutineContext,
             httpClientEngine = httpClientEngine,
             httpClientConfig = httpClientConfig,
         )
@@ -85,7 +86,7 @@ class MatrixClientServerApiClientImpl(
     syncBatchTokenStore: SyncBatchTokenStore = SyncBatchTokenStore.inMemory(),
     syncLoopDelay: Duration = 2.seconds,
     syncLoopErrorDelay: Duration = 5.seconds,
-    syncCoroutineScope: CoroutineScope? = null,
+    coroutineContext: CoroutineContext = Dispatchers.Default,
     httpClientEngine: HttpClientEngine? = null,
     httpClientConfig: (HttpClientConfig<*>.() -> Unit)? = null,
 ) : MatrixClientServerApiClient {
@@ -101,8 +102,8 @@ class MatrixClientServerApiClientImpl(
     private val coroutineExceptionHandler = CoroutineExceptionHandler { _, exception ->
         log.error(exception) { "There was an unexpected exception in sync. This should never happen!!!" }
     }
-    private val finalSyncCoroutineScope: CoroutineScope =
-        syncCoroutineScope ?: CoroutineScope(coroutineExceptionHandler)
+    private val coroutineScope: CoroutineScope =
+        CoroutineScope(coroutineContext + SupervisorJob(coroutineContext[Job]) + coroutineExceptionHandler)
 
     override val appservice: AppserviceApiClient = AppserviceApiClientImpl(baseClient)
     override val authentication = AuthenticationApiClientImpl(baseClient)
@@ -112,7 +113,7 @@ class MatrixClientServerApiClientImpl(
     override val room = RoomApiClientImpl(baseClient, eventContentSerializerMappings)
     override val sync = SyncApiClientImpl(
         baseClient,
-        finalSyncCoroutineScope,
+        coroutineScope,
         syncBatchTokenStore,
         syncLoopDelay,
         syncLoopErrorDelay
@@ -123,7 +124,13 @@ class MatrixClientServerApiClientImpl(
     override val push = PushApiClientImpl(baseClient)
 
     override fun close() {
-        finalSyncCoroutineScope.cancel()
+        coroutineScope.cancel()
         baseClient.close()
+    }
+
+    override suspend fun closeSuspending() {
+        val job = coroutineScope.coroutineContext.job
+        close()
+        job.join()
     }
 }