+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Added tests
+- Introduced `MatrixClient::closeSuspending`
+- Introduced `MatrixClientServerApiClient::closeSuspending`
+
 ### Deprecated
 
 ### Removed